Once again changes many things
- `s` as `ge` bind
- treehopper
- gruvbox-baby but implemented with nightfox
- toggle lsp-sig
- filename with folder context in lualine
- cool tab terminal with <Leader><S-Enter>
- treesitter movement for comments

+++ lua/core/util.lua
@@ -24,7 +24,7 @@ U.string_starts = function(string, start)
     return string.sub(string, 1, string.len(start)) == start
 end
 
-U.small_border = "double" -- smaller helper floats
+U.small_border = "single" -- smaller helper floats
 U.big_border = "shadow" -- ui window floats
 
 return U
